+++ DataLayer/EfClasses/Ingredient.cs
@@ -4,7 +4,7 @@ namespace DataLayer.EfClasses
 {
     public class Ingredient
     {
-        public Ingredient(Article article, uint quantity, Unit unit)
+        public Ingredient(Article article, double quantity, Unit unit)
         {
             Article = article;
             Quantity = quantity;
@@ -17,7 +17,7 @@ namespace DataLayer.EfClasses
 
         public virtual Article Article { get; [UsedImplicitly] private set; }
 
-        public uint Quantity { get; [UsedImplicitly] private set; }
+        public double Quantity { get; [UsedImplicitly] private set; }
 
         public virtual Unit Unit { get; [UsedImplicitly] private set; }
