+++ ShopAndEat/Models/RecipeModel.cs
@@ -8,7 +8,7 @@ namespace ShopAndEat.Models
         {
             // TODO mu88: Make this smarter
             var numberOfIngredients = 30;
-            IngredientQuantities = new uint[numberOfIngredients];
+            IngredientQuantities = new double[numberOfIngredients];
             IngredientArticleIds = new int[numberOfIngredients];
             IngredientUnitIds = new int[numberOfIngredients];
         }
@@ -17,7 +17,7 @@ namespace ShopAndEat.Models
 
         [Required] public int NumberOfDays { get; set; }
 
-        [Required] public uint[] IngredientQuantities { get; set; }
+        [Required] public double[] IngredientQuantities { get; set; }
 
         [Required] public int[] IngredientArticleIds { get; set; }
